mostly fixed Prim's, noted pivoting from A* to LP in GOALS

+++ prims.h
@@ -11,7 +11,7 @@ class Prims{
         Prims(Graph g);
         void primsMST();
         void printMST(std::map<std::string, std::string> parent);
-        std::string minKey(std::map<std::string, int> key, std::map<std::string, bool> mstSet);
+        std::string minKey(std::map<std::string, double> key, std::map<std::string, bool> mstSet);
         Graph graph;
         std::map<std::string, std::map<std::string, double>> adj_list;
         int node_count;
